Make fuzz a whole number, and reject one that is not
Nothing useful lives below 1: shifting every pixel by one unit on every channel scores 0.97, and the tightest bound under that is 0, which already means exact. So the scale needs no fractions and no percentage, and --expect-fuzz and the trailing script token now take the same kind of number, which they did not before -- the option was an int while the token was parsed with float().

The token is still read as a float, so that a script written against the old metric is told what is wrong with its number rather than having it silently read as the next command. A fractional or out-of-range one is a parse error naming what was written.

The score stays a float. A screen differing in one blue bit scores 0.44, and rounding that onto the same whole-number scale would make expect FILE 0 accept a screen that is not the target.

+++ vncdotool/imagematch.py
@@ -76,7 +76,7 @@ def fuzz_for_format(pixel_format: pixelformat.PixelFormat) -> int:
     return bound_for_channels(pixelformat.channel_tolerance(pixel_format))
 
 
-def matches(actual: Image.Image, expected: Image.Image, fuzz: float, blur: int = 0) -> bool:
+def matches(actual: Image.Image, expected: Image.Image, fuzz: int, blur: int = 0) -> bool:
     if actual.size != expected.size:
         return False
     worst = ImageStat.Stat(
